feat: move types to a specific file
