+++ packages/sidequest/src/operations/index.ts
@@ -1,3 +1,4 @@
 export * from "./job";
 export * from "./queue";
 export * from "./sidequest";
+export * from "./types";
